requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
colorama
python-dotenv
reportlab
